Switch to pregenerated JSON files for D1
