+++ package.json
@@ -123,7 +123,6 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.1.2",
-    "@destiny-item-manager/zip.js": "github:DestinyItemManager/zip.js",
     "@fortawesome/fontawesome-svg-core": "^1.2.4",
     "@fortawesome/free-brands-svg-icons": "^5.3.1",
     "@fortawesome/free-regular-svg-icons": "^5.3.1",
@@ -172,7 +171,6 @@
     "reselect": "^4.0.0",
     "rxjs": "^5.5.10",
     "simple-query-string": "^1.3.0",
-    "sql.js": "github:DestinyItemManager/sql.js",
     "textcomplete": "^0.17.1",
     "tooltip.js": "1.2.0",
     "typesafe-actions": "^3.0.0",
